styling fixes to checklist components, navigation and UserChip (fix) moved items in navigation into own NavItems for better structure.
