+++ src/components/RoutesContainer.tsx
@@ -4,7 +4,7 @@ import PageNotFound from '../pages/PageNotFound'
 import { IndexCheckLists } from '../pages/checklist'
 import { CheckList } from '../pages/checklist/allchecklists/Index'
 import { MyCheckLists } from '../pages/checklist/checkListID/Index'
-import { CompletedChecklists } from '../pages/checklist/completedChecklists/CompletedChecklist'
+import { CompletedChecklists } from '../pages/checklist/completedChecklists/Index'
 import { EditCheckListContextProvider } from '../pages/checklist/editchecklist/context/editCheckListContextProvider'
 import { EditCheckList } from '../pages/checklist/editchecklist/editCheckList'
 import { FillOutCheckList } from '../pages/checklist/fillOutChecklist'
