+++ src/components/navigation/hooks/DefaultNavigation.tsx
@@ -1,19 +1,8 @@
-import { Icon, Typography } from '@equinor/eds-core-react'
-import { assignment, checkbox, lock } from '@equinor/eds-icons'
 import { action } from '@storybook/addon-actions'
 import { useState } from 'react'
 import { Link } from 'react-router-dom'
-import {
-    ActiveChecklistContainer,
-    ChecklistContainer,
-    FooterContainer,
-    ImageContainer,
-    ImageContainerActive,
-    StyledList,
-    StyledTab,
-    StyledTabs,
-    ThirdTab,
-} from '../styles'
+import { FooterContainer, StyledList, StyledTab, StyledTabs } from '../styles'
+import { Items } from './NavItems'
 
 export const DefaultNavigation: React.FC<{ hideNavbar: boolean }> = ({
     hideNavbar,
@@ -31,59 +20,6 @@ export const DefaultNavigation: React.FC<{ hideNavbar: boolean }> = ({
         action('handleBlur')(e.target.textContent)
     }
 
-    const items = [
-        {
-            name: 'Tab 1',
-            value: (index?: number) =>
-                activeTab === index ? (
-                    <ImageContainerActive>
-                        <Icon data={checkbox} size={24} />
-                        <Typography variant="caption" color={'#73b1b5'}>
-                            Punches
-                        </Typography>{' '}
-                    </ImageContainerActive>
-                ) : (
-                    <ImageContainer>
-                        <Icon data={checkbox} size={24} />
-                        <Typography variant="caption" color={'white'}>
-                            Punches
-                        </Typography>{' '}
-                    </ImageContainer>
-                ),
-        },
-        {
-            name: 'Tab 2',
-            value: (index?: number) =>
-                activeTab === index ? (
-                    <ActiveChecklistContainer>
-                        <Icon data={assignment} size={24} />
-                        <Typography variant="caption" color={'#73b1b5'}>
-                            Checklists
-                        </Typography>
-                    </ActiveChecklistContainer>
-                ) : (
-                    <ChecklistContainer>
-                        <Icon data={assignment} size={24} />
-                        <Typography variant="caption" color={'white'}>
-                            Checklists
-                        </Typography>
-                    </ChecklistContainer>
-                ),
-        },
-        {
-            name: 'Tab 3',
-            value: () => (
-                <ThirdTab>
-                    {' '}
-                    <Icon data={lock} size={24} />{' '}
-                    <Typography variant="caption" color={'white'}>
-                        ---
-                    </Typography>
-                </ThirdTab>
-            ),
-        },
-    ]
-
     return (
         <FooterContainer>
             {!hideNavbar && (
@@ -95,13 +31,17 @@ export const DefaultNavigation: React.FC<{ hideNavbar: boolean }> = ({
                     onBlur={handleBlur}
                 >
                     <StyledList>
-                        {items.map(({ name, value }, index) => (
+                        {Items({ activeTab }).map(({ name, value }, index) => (
                             <StyledTab
                                 as={Link}
                                 to={
                                     name === 'Tab 1'
                                         ? '/ListPunches'
-                                        : '/CheckList'
+                                        : name === 'Tab 2'
+                                        ? '/CheckList'
+                                        : name === 'Tab 3'
+                                        ? '/404'
+                                        : ''
                                 }
                                 key={name}
                             >
