+++ src/pages/checklist/allchecklists/UserChip.tsx
@@ -1,7 +1,6 @@
 import { Icon, Typography } from '@equinor/eds-core-react'
 import { assignment_user } from '@equinor/eds-icons'
 import { FC } from 'react'
-import { formatDate } from '../../../Helpers'
 import { AllWorkFlows } from '../workflow/types'
 import { useUserContext } from './../../../pages/users/context/userContextProvider'
 import { StyledChip } from './styles'
@@ -14,7 +13,7 @@ export const UserChip: FC<UserChip> = ({ workflow }) => {
     const { currentUser } = useUserContext()
 
     return (
-        <>
+        <div>
             <StyledChip variant="default">
                 <Icon data={assignment_user} color="#243746" />
                 <Typography
@@ -35,15 +34,6 @@ export const UserChip: FC<UserChip> = ({ workflow }) => {
                     )}
                 </Typography>
             </StyledChip>
-            <Typography
-                variant="caption"
-                token={{
-                    textAlign: 'left',
-                    fontSize: '0.7rem',
-                }}
-            >
-                {formatDate(workflow.createdDate)}
-            </Typography>
-        </>
+        </div>
     )
 }
